remove multiouch from core
