+++ lib/appium_lib/core/core.rb
@@ -6,7 +6,6 @@ require_relative 'driver'
 
 # for multi touch related methods
 require_relative 'device/touch_actions'
-require_relative 'device/multi_touch'
 
 require_relative 'android'
 require_relative 'android_uiautomator2'
