+++ lib/appium_lib/core/ios.rb
@@ -1,7 +1,7 @@
 # loaded in common/driver.rb
 require_relative 'ios/search_context'
 require_relative 'ios/device'
+require_relative 'ios/touch'
 
 require_relative 'ios/uiautomation/patch'
 require_relative 'ios/uiautomation/bridge'
-require_relative 'ios/uiautomation/touch'
