+++ lib/appium_lib/core/ios/uiautomation/bridge.rb
@@ -11,8 +11,7 @@ module Appium
 
             Core::Ios::Uiautomation.patch_webdriver_element
 
-            ::Appium::Core::Ios::Touch.extend_touch_actions
-            ::Appium::Core::Ios::Touch.extend_multi_touch
+            Core::Ios::Touch.extend_touch_actions
           end
         end
       end
