+++ lib/appium_lib/core/ios/xcuitest/bridge.rb
@@ -11,8 +11,7 @@ module Appium
             target.extend Appium::Ios::Device
             target.extend Appium::Ios::Xcuitest::Device
 
-            ::Appium::Core::Ios::Touch.extend_touch_actions
-            ::Appium::Core::Ios::Touch.extend_multi_touch
+            Core::Ios::Touch.extend_touch_actions
           end
         end
       end
