+++ lib/appium_lib/core/ios_xcuitest.rb
@@ -1,8 +1,9 @@
 # loaded in common/driver.rb
 require_relative 'ios/search_context'
 require_relative 'ios/device'
+require_relative 'ios/touch'
 
 require_relative 'ios/xcuitest/search_context'
 require_relative 'ios/xcuitest/device'
 require_relative 'ios/xcuitest/bridge'
-require_relative 'ios/xcuitest/touch'
+
